Tag a new version of compose.yml on auto-update
When a new version of the service image gets added to the compose.yml
automatically then tag the compose.yml with a new version as well.
The service image is only one dependency (or library, if you will) that
the compose.yml uses and therefore the versions of the high-level script
are tracked independently of the versions of software it uses.

Issue #8 Add (or update) action to tag new versions of the compose.yml