Update images and environment variables
Declaring HOST=0.0.0.0 allows the reverse proxy to reach the web service
successfully. Passing through other environment variables both to the
web and database services related to database settings should help web
connect to the database instance.

The reason this is connected to Issue #9 is it is a small change to the
compose script in order to make a new tag to test out deployment.

Issue #9 Add (or update) action to send a tag version to a machine