Add an auth container, namely keycloak
Add an authentication and authorization container to the PDC services
and integrate it with the existing database service and reverse proxy
service.

This commit does not yet integrate with the web service. The purpose
of pushing this is to allow testing of the changes thus far.

The bitnami postgres image seems streamlined for a single database but
adding a second database is done in a script keycloakInitDb.sh. If one
needs to have the script run again, remove a file that the bitnami
image create after run called .user_scripts_initialized.

See environment variable use in both compose.yml and keycloakInitDb.sh
and the updated list in .env.example.

Issue #49 Add authentication/authorization container, integrate it