Enable gzip HTTP compression for JSON data
After consideration of the [BREACH attack](https://breachattack.com)
we enabled HTTP compression. The reason we do not think the back-end
service is vulnerable to BREACH is the following. Even though some
query or URL parameter data are reflected in the JSON body, all
protected or secret data require a JSON Web Token auth header. When
there is no valid authentication in the HTTP headers (which are not
compressed), the response does not reflect query or URL parameter
data. Therefore the attack described should not be possible without
first compromising the credentials of a user at which point the attack
would be pointless because the attacker had already gained access to
protected or secret data by using the victim's credentials.

Issue #84 Compress API response