Add a dependabot configuration
Hopefully dependabot will "just work" but the presence of
https://github.com/dependabot/dependabot-core/issues/390 suggests it
will not. This repository doesn't have a `Dockerfile`, only a
`compose.yml`.

As of this commit it is not expected that dependabot will immediately
work for Docker images, but we hope the above issue will be fixed and
at that point dependabot will report new images.