Restart the reverse proxy container
Previous attempts to send the targeted `reload` signal did not let
nginx see the freshly-installed certificates. The workaround was to
restart the whole container. Therefore this change incorporates that
workaround and restarts the whole container every time. Nginx restarts
pretty quickly so it is not a lot of down time. The downside might be
interrupted requests during the restart.

Fixes #90 Fresh certificates should be presented by nginx