Remove auth server cert renewals
The authentication and authorization server is getting split off onto
its own host. As of this writing, the test auth server is separated.
This change removes the attempts to renew certificates for the auth
service now that it is going to be on a different machine.

See #94 Run auth service on a separate machine