Expose the deployed software version via nginx
Before this change, there was no way to tell what version of the
software is currently running on the system. This made it impossible
for a CI script, for example, to ask which version is currently there.
With this change, a CI script (e.g. in a GH action) can ask which
version is deployed on the test environment, verify that a page or two
return HTTP 200 (besides this one), and then auto-deploy to production
with some confidence that the production deployment will succeed.

This removes one of the objections to the current deployment pipeline
and also removes the objection to the reason we do not yet have
automatic deployments to production: the test auto-deployment should
succeed prior to a production auto-deployment.