Auto-deploy to production after test deployment
Without this change, someone needs to verify that the automated
deployment to the test environment succeeded and then manually type
two or three commands to deploy to production. At the moment, there is
only one person who has ever done such production deployments.

With this change, however, the deployment to the test environment gets
verified by visiting a URL that exposes the current deployed version.
When the body of a request sent to that URL returns the same string as
the version (tag) sent to this deployment job, this action can safely
conclude that the test deployment has succeeded. How and why can it
make such a haughty conclusion? Several reasons:

* The `deploy.sh` script saves the new version only on success.
* The server in the `reverse-proxy` container serves the version that
  the `deploy.sh` script wrote.
* The `reverse-proxy` container does not start until the `database`
  and `web` containers are running according to docker health checks.

In other words, accessing this version string implies that the state
of the environment is OK. If the version string matches in the test
environment, this gives some confidence that auto-deploying to the
production environment is relatively safe. Therefore it auto-deploys
to production. This means that merging pull requests in the `service`
repository is sufficient to cause automatic deployments to production
and at the same time the production deployment is safe due to the same
steps having run for a deployment to test. It means no more asking for
some person to deploy to production anymore!

Since "send tag to machine" is indecipherable, the action is renamed.

Issue #106 Auto-deploy to production