Pass env vars appropriately to deployment CI jobs
Before this commit, the `KNOWN_HOSTS` variable was not sent to
`trigger_deployment.sh`, causing failures. It also seems inappropriate
to pass secrets to 3rd-party actions if avoidable, so this commit also
removes access to the secrets for the action that checks for the given
tag to be on the main branch.

Issue #106 Auto-deploy to production