Expect a line terminator in software-version URL
Without this change, the CI action that asks which version is running
in a target environment would fail because of a line terminator
difference in the expected string versus the returned string.

This change attempts to fix the issue by using a regular expression
that allows trailing whitespace. An attempt to add an escaped endline
to the expected response body fails because inputs are trimmed in the
action (at the time of this commit).

Issue #106 Auto-deploy to production